requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
faiss-cpu
rank-bm25
numpy
streamlit